Refactor forms and remove unused items
